Fix the type of struct bbf's arc_count member: according to gcc's definition of struct bb_function_info in libgcc2.c, it should be an int, not a long.

+++ usr.sbin/kernbb/kernbb.c
@@ -29,7 +29,7 @@ typedef long long gcov_type;
 
 struct bbf {
 	long	checksum;
-	long	arc_count;
+	int	arc_count;
 	u_long	name;
 };
